[DF-8266] build fixed, warning fixed
Former-commit-id: 08e188cd32309cd3bd67a61d74e31ff9d1e53267

+++ Tools/QuickEd/Classes/UI/mainwindow.cpp
@@ -525,7 +525,7 @@ void MainWindow::OnGlobalClassesChanged(const QString &str)
 
 void MainWindow::OnLogOutput(Logger::eLogLevel logLevel, const QByteArray &output)
 {
-    if((1 << logLevel) & acceptableLoggerFlags)
+    if(static_cast<int32>(1 << logLevel) & acceptableLoggerFlags)
     {
         logWidget->AddMessage(logLevel, output);
     }
